Add types.ts to pretest
