+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 npm-debug.log
 yarn.lock
+types.js
